PatinaPkg: Add measurement bit note in PatinaPerformanceConfig.h
Adds a note that points to the Patina Performance component
readme to show how bits are configured in the measurement mask.

The actual enum definitions is in the `patina` crate, however,
the Patina Performance config definition is more relevant here
and might track movement of that enum in the future to a different
location.

The Microsoft open source copyright does not have the text
"All rights reserved" included so that is fixed. This is the only
file in the repo found to have that text so the update is simply
included in this commit.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>