MpDxe: Prevent override of the BSP GDT configuraiton
Currently MpDxe will overrite the GDT for all cores to setup exception
handler stacks based on the platform configuration. This causes issues
for Patina where the BSP exception handlers are more robust and require
a larger stack then the EDK2 code accounts for.

This change moves the BSP handler for the function  to be a no-op so
that the BSP GDT configuration is preserved.